end commit

+++ src/Component.java
@@ -1,15 +1,15 @@
 public class Component {
     public String name;
-    public float price;
+    public double price;
     public String description;
 
-    public Component(String name, float price, String description) {
+    public Component(String name, double price, String description) {
         this.name = name;
         this.price = price;
         this.description = description;
     }
 
-    public float getPrice() {
+    public double getPrice() {
         return price;
     }
 }
